Feat: deleted hanbang & koryo and added skeleton in hanjin, nh, postbox, cupost
